[riscv] Use pc_offset_for_safepoint instead of blocking trampolines
We currently block trampolines for one instruction following certain
jumps. This means that we have to insert nops in various places if we
need to have the trampolines emitted - and it makes it hard to guarantee
that we can always emit trampolines before large constant pool sections.

With this change, we transition to recording the safepoint pc right
after emitting the call instruction (jal/jalr) like it is done on the
MIPS and LOONG architectures. The safepoint recording uses the correct
pc by calling pc_offset_for_safepoint() instead of using pc_offset().

Change-Id: I2307c9ec08143f0be1e629fcbe59c0e60dc261d9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6652837
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#100986}