[sandbox][api] Support memory protection keys in VirtualAddressSpace
This CL adds initial support for memory protection keys to the
VirtualAddressSpace API. In particular, it is now possible to allocate a
subspace using a given key, in which case the space will ensure that all
pages allocated through it use this key.

The problem this is solving is the following: previously, we were simply
setting the sandbox pkey on the sandbox address space at initialization
time. This works fine for the V8 heap (at the start of the sandbox), as
that will only use mprotect to allocate and free memory. However, other
parts of the sandbox address space manage memory differently: there,
mmap with MAP_FIXED is used to replace existing pages with "fresh" pages
(see VirtualAddressSubspace::FreePages). This, however, resets the pkey
under the hood as it's effectively an atomic munmap + mmap. As such, the
proper solution to this is to make the VirtualAddress(Sub)space class
aware of memory protection keys so it can ensure that all pages
allocated through it have the correct key set.

This should be a nice and robust way to integrate memory protection keys
into our memory management APIs. As a follow-up, we could also add an
optional memory protection key parameter for the regular page allocation
routines and use it instead of MemoryProtectionKey::SetPermissionAndKey.
That way, the MemoryProtectionKey API would become responsible only for
managing the keys, and the VirtualAddressSpace API would manage the
memory pages that use these keys.

Bug: 350324877, 40643847, 416209124
Change-Id: I5c784741b7458ce56d862ac2b5ed8dbc6fa7b1e1
Cq-Include-Trybots: luci.v8.try:v8_linux64_pku_dbg,v8_linux64_pku_rel
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6657446
Reviewed-by: Stephen Röttger <sroettger@google.com>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100982}