[builtins] Pass the real context in a few runtime calls
Creating new Error objects needs the context to be set in order to load
the error function object. A few runtime calls in Intl builtins did not
pass a context; these are fixed in this CL.

The story is more complex in optimizing compilers, where we'd have to
introduce exceptional control flow to properly handle errors. For these,
we now ensure that Factory::NewError crashes consistently on a CHECK
instead of throwing (and we still pass in the EmptyContext/smi zero).

Drive-by: Remove dead code.

Fixed: 424310766
Change-Id: Ieaca62cf53c0898c0c502cf3579e3f0892264b5b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6646505
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100995}