Revert temporary fix for issue 40070746
Now that we have the proper fix in place, we no longer need massive
guard regions after the sandbox to mitigate the issue. However, as
described in the issue, the previous guard region size of 32GB is not
quite sufficient, so we now use 64GB to be on the safe side.

Bug: 40070746
Change-Id: Ib4fe31e21bab6442a279938500878703d178c16e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6140873
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Auto-Submit: Samuel Groß <saelo@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#100998}