[riscv] Remove rare, but broken optimization: auipc; jalr -> jal; nop
The optimization moved the jump instruction 4 bytes back, thus changing
the pc saved in the ra register. This turned out to be problematic in
the presence of a safepoint associated with the original return address
that would no longer match the pc saved in the frame.

Change-Id: Ib87e37e9a6b10effef7f0c551e437ccfabe6d55d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6668653
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101017}