[riscv] Use BlockTrampolinePoolScope in more places
This is the next step towards getting rid of BlockTrampolinePoolFor. To
make this robust, I've started tracking the amount of code we generate
while the pool is blocked. Unfortunately, we currently cannot insist on
not generating more than that, because we may end up emitting a constant
pool section of an unforeseeable size while blocked.

We now save an unused entry in the trampoline pool, because we can avoid
using it for the jump around the pool which will always be a near jump.

Change-Id: I90925b0bd8b8662aa37216c297bd3c16302a0850
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6664461
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#101041}