[runtime] Introduce kClearedWeakValue and cleanup ClearedValue()
It used to be important that the upper part of a non-Smi value is equal
to respective cage base but since https://crrev.com/c/6402513 it's not
the case anymore.
This CL relaxes the DCHECK in CompressObject() to allow baseless
cleared weak values.

Although we can simply use kClearedWeakValue instead of ClearedValue()
or ClearedTrustedValue() for manual clearing of certain weak fields,
having these functions might still be useful to ensure that we don't
accidentally mix up cleared weak references belonging to different
cages.

Bug: 424154602
Change-Id: I03a351c2c1615455c076cf73d38e592b75bdcd54
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6673894
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101053}