[riscv] Fix shared-everything/atomic-instructions case failed in pointer compression mode.
We use lr_w to to implement AtomicCampredExchange in Pointer
Compression mode. But if object is pointer, it need decompressed. And
the expected had been decompressed. To ensure the accuracy of the
comparison results, we need to sign extend the expected. Also we can't
emit sign extend instr in instr selector, because There are have chance
to use a same register for i.InputRegister(2) and i.OutputRegister(0).

Change-Id: Ideadf69b9b9b91ae26f816daf2ff74c00df5a5bd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6671250
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Commit-Queue: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101039}