Unify error handling macros for different Maybe types
Suggested by jgruber@; We have a distinction between Maybe and
MaybeHandle/MaybeDirectHandle when it comes to the error handling
macros. We do not need such a distinction as long as they have similar
interfaces, which is easy to do.

This doesn't actually switch the macros over: I imagine that will be
a highly bitrotty CL, but it makes the macros proxy each other, making
it easy to do a find-and-replace later.

Open to suggestions on whether the new To/IsEmpty APIs should be
considered shims or if we should just move the old users wholesale onto
them. Open to either.

Bug: 427539322
Change-Id: I6a6a6964ef9cd655319f198da4ea62117ba87eb7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6678531
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Manish Goregaokar <manishearth@google.com>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Auto-Submit: Manish Goregaokar <manishearth@google.com>
Cr-Commit-Position: refs/heads/main@{#101122}