[heap] Make --stress-compaction more aggressive
So far --stress-compaction only selected even-indexed pages as
evacuation candidates. However we should be able to choose all pages
as evacuation candidates as well. This might make it easier to
reproduce a crash if the object deterministically falls on an odd
page at a specific GC.

Change-Id: Id1d8fa0d529fce94af0974d29f6bf0ec58d06d4d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6678671
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101048}