[wasm][growable-stacks] Fix IsOnCentralStack check
When computing the central stack bounds in Isolate::IsOnCentralStack()
check, we incorrectly use the secondary stack limit margin instead of
the central stack margin. This is the margin that we allow C++ to
overflow the JS stack limit by, and it is stricter for secondary stacks
than for the central stack. So this can lead to false positives when C++
code runs near the stack limit.

R=clemensb@chromium.org

Fixed: 425827679
Change-Id: I65614346d82848bce4f3abd95e3e3b730fdce5c2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6678674
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101059}