[maglev] Fix merge states in polymorphic calls
We were fixing the predecessor count (if a continuation was dead) but we
might already have created a MergePointInterpreterFrameState with the
wrong predecessor count.

This CL simplifies dead bytecode detection so that it's not done inside
VisitSingleBytecode but outside. This way callers can choose to not
detect and handle dead bytecodes, which is the right thing to do when
only one polymorphic branch is dead.

Fixed: 426157225
Bug: 411351177
Change-Id: I45a785d4441c6216b176e964a5c1eaea305e5968
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6658623
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101061}