[riscv] Move F32x4{Min|Max} and F64x2{Min|Max} to code generator
Moves the logic of the F32x4{Min|Max} and F64x2{Min|Max} visit-functions
from the instruction-selector to the code-generator.

The visit function were using the `kSimd128ScratchReg`, which isn't
supposed to be used at this level. A debug check caught the usage of it,
leading to failures in debug builds.

Change-Id: Ic4482fadf28f3318deea3f227faa10c65116ebff
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6674274
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Commit-Queue: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101074}