[riscv] Block the constant pool when blocking the trampoline pool
We must be able to generate the trampoline pool before we emit a large
constant pool. Otherwise, we will run into situations where the short
branches cannot reach the other side of the constant pool. This change
fixes that by blocking the constant pool generation whenever we block
the trampoline pool.

This also removes support for blocking for a number of instructions,
since this is now unused.

Finally, this fixes a bug in Maglev's handling of deoptimization exits
on RISC-V. We now avoid getting the branch trampoline pool emitted in
the middle of the section with the exits - and we no longer generate
unused shared tail calls to the deoptimization builtins.

Change-Id: Ie9bdf6596e2f031481b40d36c35ec782346b0e1e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6678572
Commit-Queue: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101073}