[riscv] Move the logic of Simd convert functions to the code generator
Moves the logic of the following visit functions from the
instruction-selector to the code-generator:
- I32x4SConvertI16x8Low
- I32x4UConvertI16x8Low
- I16x8SConvertI8x16High
- I16x8SConvertI32x4
- I16x8UConvertI32x4
- I8x16SConvertI16x8
- I8x16UConvertI16x8
- I32x4SConvertI16x8High
- I32x4UConvertI16x8High
- I16x8SConvertI8x16Low
- I16x8UConvertI8x16High
- I16x8UConvertI8x16Low

The visit functions were using the `kSimd128ScratchReg`, which isn't
supposed to be used at this level. A debug check caught the usage of it,
leading to failures in debug builds.

Change-Id: Ibbce7c584d2b27386227bff6566f7f992144870c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6673636
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#101084}