[riscv] Move RoundingAverageU logic to code generator
Moves the logic of the I16x8RoundingAverageU and
I8x16RoundingAverageU visit functions from the instruction-selector
to the code-generator.

The I8x16RoundingAverageU function was using the
`kSimd128ScratchReg` register, which isn't supposed to
be used at this level. A debug check caught the usage of it, leading to
failures in debug builds.

The I16x8RoundingAverageU function could have stayed the way it was, but
it's cleaner to have them both handled the same way.

Change-Id: I1cb5d6db41da4eb9066bf2a901a654a5f2816994
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6675034
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#101087}