[riscv] Move Dot logic to code generator
Moves the logic of the following functions from the instruction selector
to the code generator:
- I32x4DotI16x8S
- I16x8DotI8x16I7x16S
- I32x4DotI8x16I7x16AddS

The I32x4DotI8x16I7x16AddS function was using the
`kSimd128ScratchReg` register, which isn't supposed to
be used at this level. A debug check caught the usage of it, leading to
failures in debug builds.

The other Dot functions could have stayed the way they were, but it's
cleaner to have them all be handled the same way.

Change-Id: Ifcae2186823112f5aa11c2b633c43f31a4e27ac2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6675054
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#101088}