[wasm-fuzzing] Fix FunctionSig lifetime in random module generation
Replace temporary FixedSizeSignature with CreateSignature for main
function signature to ensure proper memory management when exporting
modules.

The previous code created a stack-local FixedSizeSignature that would
be destroyed at the end of the function scope, but the signature
pointer could be stored in module_exports which lives beyond the
function scope. When modules are exported and their signatures accessed
later, this would result in use-after-free.

This change ensures the signature is properly allocated in the zone
and has the correct lifetime for exported modules.

Change-Id: I6ee6c419be7ea1d165b9dc3880f99d0a68e82f28
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6681576
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Emmanuel Romero Ruiz <emromero@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#101112}