[bigint][sandbox] Harden FromString against concurrent mutation
The {FromStringLarge} implementation uses the allocated result as
temporary storage at first. This algorithm crucially relies on that
result having the right size. We do allocate it with the right size,
but a concurrent mutator could immediately muck with it; so guard
against that with a CHECK.

Fixed: 427600180
Change-Id: I19366ae8562aefd21af1d850df4155c016cdb904
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6680320
Reviewed-by: Samuel Groß <saelo@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101119}