Further separate the two purposes of GetStringOption
GetStringOption has two overloads: one which fetches a string (and tells
you if it succeeded), and one which matches a string option against a
set of values. The first overload does double-duty: it also checks
membership in a list of options if provided, but doesn't if it is empty.

This is inefficient: the match-string-option path ends up iterating
through options twice, and furthermore the callers that use the
fetch-string-option path end up converting to a std::string anyway.

In the medium term these should probably be called GetStringOption and
MatchStringOption.

Bug: 415359720
Change-Id: I6a6a69646d7bedcb2a45748c16d27d6eab751f08
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6685912
Reviewed-by: Rezvan Mahdavi Hezaveh <rezvan@chromium.org>
Commit-Queue: Manish Goregaokar <manishearth@google.com>
Cr-Commit-Position: refs/heads/main@{#101146}