Correctly handle REQUIRED case in GetStringOption
and use the "match options" version of GSO more.

GSO has two overloads: a "fetch option" one that checks that the option
is one among many possible values, and a templated "match option" one
that matches amongst them and returns a corresponding enum.

This adds handling for the REQUIRED case to the "fetch option" one, and
makes it usable in more cases.

Then, in cases where the "fetch option" version is used where the goal
is to match on an option, this CL changes it to using the "match option"
one, which is faster (avoids performing ToStdString conversions).

Bug: 415359720
Change-Id: I6a6a696429794fe031c0308a36e236c81dd9506c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6685911
Commit-Queue: Manish Goregaokar <manishearth@google.com>
Reviewed-by: Rezvan Mahdavi Hezaveh <rezvan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101145}