[heap] Set pointer cage base when freeing pool pages.
On multi-cage configs we need to be sure the pointer
cage base and current sandbox are set when the pool
pages are freed.

Without this change we fail in
MemoryChunk::MetadataTableIndex which uses
IsolateGroup::current and
V8HeapCompressionScheme::base().  This is called
from the MemoryChunkMetadata destructor, called
from DeleteMemoryChunk.

Change-Id: Iaef0d96063524ca0ba27d4bddf4251b3148f664b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6687810
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101132}