[base] Port base::FunctionRef from Chromium
absl::FunctionRef is banned in Chromium because it is too loose with
implicit conversions. base::FunctionRef is a wrapper around it which
checks for and forbids these conversions. This wrapper would be useful
in V8 too, so port it to V8 along with the relevant helper classes.

Change-Id: Ib67e1b7fe6445860fbcc0dad092b9a4f427bb433
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6689117
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101140}