[sandbox] Introduce --strict-pkey-sandbox
When this (experimental) feature is enabled, and hardware sandboxing
support is available, we will remove write access to the default pkey
(key zero) in sandboxed execution mode. As such, in this mode no
out-of-sandbox memory can be written to with the exception of "sandbox
extension" memory (marked as such through the extension_pkey_).

This mode can (currently) only be used on x64 Linux and requires some
workarounds (see flag description). It is also not yet fully functional
as there are some remaining code fragments that run in sandboxed
execution mode but write outside of the sandbox. These will be fixed in
follow-up CLs.

Once this mode is fully functional, we should make this the default mode
for sandbox hardware support and remove the non-strict mode (in which
out-of-sandbox memory must manually be tagged with a pkey).

Bug: 350324877
Change-Id: I6eccca3c0e825c4e20bd32f5ee8a814f86040cf5
Cq-Include-Trybots: luci.v8.try:v8_linux64_pku_dbg,v8_linux64_pku_rel
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6665100
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101137}