[riscv][liftoff] Use safepoint pc after a call
Some architectures, like RISCV, may emit a constant-pool or trampolines
after a call. When the compiler wants to have the return address of
a call it needs to use 'pc_offset_for_safepoint' which is set to the PC
after the call.

This fixes the wasm-step-a-lot test on RISCV. It had a trampoline
inserted after 189 WASM instructions, which broke the wasm-debugger, as
it couldn't find the debug information anymore.

Change-Id: I6a361ec1924bced22a49b444a83c404a32a912b5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6680273
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101182}