[api] Inline HandleScope operations.
The handle operations are very performance critical.
(HandleScope constructor/destructor, Local<T> creation).
Internally they are already inlined, but for API users they
turn into calls.  This change duplicates some of the functions
from handles/handles-inl.h into the include directory so that
the most critical operations can be inlined.

There was an assert in the HandleScope constructor that used
IsLockedByCurrentThread which is a fairly expensive call,
at least on Linux where it does a syscall for gettid().  This
has been moved behind V8_ENABLE_CHECKS so it is not active
in release mode.

Change-Id: I39a0c5f677be7d5b6f374f2b94df712499b7e174
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6687169
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101166}