Remove g_thread_in_wasm_code flag
Keeping track of whether a thread is currently executing Wasm code or
not has become quite complex (see size of this change) and therefore a
maintenance burden. Furthermore, this feature now conflicts with future
V8 sandbox advancements as we want to forbid writes from sandboxed code
(which includes all Wasm code) to out-of-sandbox memory. As such, we can
no longer write to thread-local storage (where the flag was stored).

This CL replaces the flag with a simpler guarding mechanism that ensures
that the trap handler is robust against nested faults and deadlocks. In
essence, we now set a thread-local flag (only) when entering the trap
handler or taking one of the locks used by it. If the flag is already
set when entering the trap handler, then we bail out immediately as it
means that we're either handling a nested fault (which we shouldn't) or
risk running into a deadlock because the current thread holds one of the
locks required by the trap handler.

Bug: 350324877, 40192807
Change-Id: Idb9fce83ac9b46ab04fd93d5bdbd8cb8c0722488
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6650680
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Mark Seaborn <mseaborn@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101213}