[wasm] Add the shared bit to type hashes
The TODO was confusing because the `exact` bit was already added. What
was missing is the `shared` bit.

This was OK since the shared bit is determined from the index, but this
made the code for hashing a bit different from the code for equality
checks. The new code make the two pretty similar again.

R=jkummerow@chromium.org

Change-Id: I1e121fbc3336998dcc9b69353730a9dab8607c7b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6687171
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101217}