[sandbox] Mark JSToWasmWrapper temp buffer as sandbox-extension memory
These buffers need to be accessible to code running in sandboxed
execution mode, in particular the generic JSToWasmWrapper. In principle,
the buffer is very similar to stack memory in that it's unsafe to have
it inside the sandbox itself as it would allow escaping from the
software-only sandbox, but is safe to mark it as sandbox-extension
memory since it will only ever be used by code running in sandboxed
execution mode. In fact, we allocate this buffer also on the stack in
case there are fewer than 10 parameters. As such, marking it as
sandbox-extension memory (but keeping it outside the sandbox) is
probably the right thing to do. With this CL, we now simply mark the
OS page on which the buffer lives as sandbox-extension memory. This is
not great as it will make other objects located on the same page
accessible, but it's probably ok as an initial workaround. In the
future, we should either use dedicated OS pages or create a
SandboxExtensionAlloc for this and similar cases (if there are any).

Bug: 429341650, 350324877
Change-Id: I989b260195404ec645a495a5e2ba6869d9c47e41
Cq-Include-Trybots: luci.v8.try:v8_linux64_pku_dbg,v8_linux64_pku_rel
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6702630
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101228}