[asmjs] Copy input before parsing
The asm parser often rewinds the scanner to re-parse the input. This is
done in particular for switches and for-loops.
To fix this issue once and for all, we now copy all characters into a
vector of utf16 values and the `AsmScanner` works on that vector.

We introduce a few helper functions in `AsmScanner` to make checking and
consuming the next character easier. This actually made some scanning
code significantly simpler.

Drive-by: Put fields last in the `AsmScanner` class (see style guide).

R=jkummerow@chromium.org

Bug: 427918760
Change-Id: I7feae8fe04f2a0f317445f553c2f08615ffbd6e4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6701450
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101227}