Wasm-in-JS-inlining: Ensure consistent module
The `JSWasmCallParameters` struct stores both a `NativeModule*` and a
`Module*`.
If they ever become inconsistent (the `Module*` not belonging to the
`NativeModule*`) we would get all kinds of problems; also it's super
cheap to get the `Module*` from the `NativeModule*`.

This CL thus removes the `Module*` and the similarly redundant signature
pointer.

To be on the safe side we now also enforce that the `NativeModule` of
all inlined Wasm functions is the same, not just the `Module`. Note that
there is typically a 1:1 relation between them, except if the native
module cache is explicitly disabled.

R=dlehmann@chromium.org

Change-Id: Ie6a9062281d58c8c513d77a06a48c11c0c85e498
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6616962
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101225}