[sandbox] Don't use signal handlers on older kernels ...
... when hardware sandboxing is active. This is a workaround for an
issue in older kernels when delivering signals on an alternate stack on
a thread that doesn't have write access to the default pkey. The
behavior has been fixed on kernel 6.12, so with this CL, we now check
for the current kernel version and disable trap handlers (such as the
Wasm trap handler) if the kernel version is too old and we want to use
hardware sandboxing.

Bug: 429173713, 350324877
Change-Id: I4a0ed92305f3da84c0b6939ad90d8e22d6ffb36b
Cq-Include-Trybots: luci.v8.try:v8_linux64_pku_dbg,v8_linux64_pku_rel
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6703657
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101234}