[sandbox] Print note for read accesses for sandbox testing
Read accesses are technically not sandbox violations. Often a read can
be turned into a write though or a read is just a byproduct of a
write.

The CL highlights that the access was a read but still treats it as
sandbox violation. Such crashes require more investigation.

Bug: 329781444
Change-Id: I443852410f367b2baf79cd11906c1e4e63769a01
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6700238
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101236}