[wasm-custom-desc] Fix "exact" casts for custom descriptors
When performing a cast to an exact static type that has a custom
descriptor, the RTT to compare against is not the object's map but
rather the last entry in its list of supertype RTTs. This is a new
third case in our zoo of type check variants. To make it work with
with over-allocating supertype lists to kMinimumSupertypeArraySize,
we need to fill the over-allocated slots with copies of the last
actual supertype (rather than `undefined`).

Fixed: 429001362
Bug: 403372470
Change-Id: I2a2c0f2f1617be9fbff5f4455b63ddc7ab3c8e6f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6700028
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101263}