[deoptimizer] Store register params first for continuation stubs
Builtin continuation stubs can have both in-register and on-stack
parameters. The intuitive way of storing these is register params
followed by stack params, since that's how they are defined in the
descriptors (the first N parameters go into registers, the remainder go
onto the stack).

Until now, the translation frames were storing stack parameters first,
and register parameters second. This had a good reason -- JS
continuation builtin frames want to have the first translation parameter
be the receiver for consistency with unoptimized frames during frame
iteration. However, this leads to awkward re-ordering of the parameters
to continuation stubs, which need to move the on stack parameters to the
start of the parameters array, and notably we were forgetting to do this
in Maglev (which caused issues in Turbolev on ia32, where there are
fewer registers and therefore more stack parameters).

We can make the parameter ordering more natural while preserving the
receiver-first property by splitting up the encoding of JS-linkage
continuation builtins and stub continuation builtins; these anyway
differ in stack parameter order (JS-linkage pushes stack arguments in
reverse order). This simplifies the stub parameter encoding (even
removing the need for an intermediate std::vector) in TF, and keeps the
Maglev one as-is.

I've also added some documentation static_asserts for the receiver-first
property of JS continuation frames, since that was surprising at first
when I tried to make _all_ register parameters be before stack
parameters, and code assuming that this is the case was spread out over
a couple of different locations.

Fixed: 428754707
Change-Id: If83d05536b0f62068be5ebb1e9ce8f5f79e6ae9a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6701447
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101281}