[heap] Emit --trace-gc-nvp log output as Perfetto events
This CL emits the log output used in --trace-gc-nvp as a perfetto
event. This allows us to access all these data from Perfetto
profiles relatively easily. For this the output is first written
into a buffer and only then printed to stdout.

This CL also adds new fields to the --trace-gc-nvp logging: 'stack'
and 'reason'. In addition this CL adds NewSpaceCapacity as a new row
in the Perfetto UI.

Bug: 408962793
Change-Id: I839321ef69387f2114575d86a26191048bb62797
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6700736
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101299}