[sandbox] Rely on death callbacks for filtering with ubsan
Ubsan uses abort() to signal issues when abort_on_error=1 which is the
Chromium fuzzer configuration. This signal is however ignored by the
sandbox testing mode to only crash on sandbox failures.  To avoid
causing problems where we run without asan but with ubsan we currently
crash on this mis-configuration. This is incompatible with e.g. just
shipping ubsan via other dependencies in production.

The CL now uses __sanitizer_set_death_callback() to set a callback in
_all_ sanitizer configurations. This callback is then used to remove
the signal handler that would filter the abort signal. This means that
when triggering an ubsan error we would remove the handler that
filters it out and should observe the crash.

Drive-by: Guard handling asan errors via __asan_report_present().

Bug: 430178746
Change-Id: I78e8d42924c90a5516a1d500900c40a48457fd46
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6714484
Reviewed-by: Samuel Groß <saelo@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101319}