[compiler] Store LoopInfos in a sorted vector
Rather than a reasonably heavyweight std::map mapping loop header
offsets to LoopInfo, store LoopInfo in a more cache-friendly
std::vector, sorted by loop header offset, and binary search for it.

Since loops are collected in a reverse bytecode walk, maintaining a
header-order sorted list would initially be quite inefficient, since
most loops would end up getting inserted close to the start of the
vector. So, we first keep a list in visitation order, and only after the
first pass of analysis completes do we sort it in header order.

Drive-by, add an `emplace` method to ZoneVector for emplacing in a
concrete location -- not used here (used in a previous version of this
patch) but still useful.

Change-Id: I9dbe9bd77eb4331a71642b96b8e6c1119310e484
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6722272
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101348}