[v8] Add flag for doing GC on backgrounded
This is part of an effort to avoid using Critical Memory Pressure to
trigger a GC in v8. One of the places this is done is when Chrome is
backgrounded, so we add a flag to v8 to trigger this separately, to
avoid regressing memory when we disconnect GC from Critical Memory
Pressure.

Bug: 430536195
Change-Id: Ibbf1bca1cfad393c4caa90ceae6725dab302d223
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6673634
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Commit-Queue: Thiabaud Engelbrecht <thiabaud@google.com>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101400}