[configs, maglev, turbofan, turbolev] Add --maglev-as-top-tier
--maglev-as-top-tier disables Turbofan and configures Maglev optimally
as the top tier (e.g., inlining more aggressively), i.e., what
--no-turbofan did before this CL.

After this CL, --no-turbofan just keeps the Maglev configuration as is
and just disables tiers beyond it.

An imperfectness remaining after this CL is that osr_from_maglev and
emitting interrupt budget checks is disabled in both configs. Ideally it
should only be disabled with --maglev-as-top-tier and enabled with just
--no-turbofan.

Change-Id: I24e11dfa0b9f318906fd2a0ce1f3130142ab8673
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6732761
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#101477}